Fix JNI bug in clientType
-- Fix icd clienttype with integer and align with generated type for clientType in other generated file
-- Fix missing semicolon in the end of ()Ljava/lang/Short

+++ src/controller/java/src/chip/devicecontroller/ICDRegistrationInfo.java
@@ -24,7 +24,7 @@ public class ICDRegistrationInfo {
   @Nullable private final Long checkInNodeId;
   @Nullable private final Long monitoredSubject;
   @Nullable private final byte[] symmetricKey;
-  @Nullable private final Short clientType;
+  @Nullable private final Integer clientType;
 
   private ICDRegistrationInfo(Builder builder) {
     this.checkInNodeId = builder.checkInNodeId;
@@ -48,7 +48,7 @@ public class ICDRegistrationInfo {
     return symmetricKey;
   }
 
-  public Short getClientType() {
+  public Integer getClientType() {
     return clientType;
   }
